+++ includes/class-wp-statistics.php
@@ -22,6 +22,7 @@ use WP_Statistics\Service\Admin\Posts\PostsManager;
 use WP_Statistics\Service\Admin\PrivacyAudit\PrivacyAuditManager;
 use WP_Statistics\Service\Admin\Referrals\ReferralsManager;
 use WP_Statistics\Service\Admin\TrackerDebugger\TrackerDebuggerManager;
+use WP_Statistics\Service\Admin\Notification\NotificationManager;
 use WP_Statistics\Service\Analytics\AnalyticsManager;
 use WP_Statistics\Service\Integrations\IntegrationsManager;
 use WP_Statistics\Service\Admin\FilterHandler\FilterManager;
@@ -166,9 +167,10 @@ final class WP_Statistics
         // Ajax area
         require_once WP_STATISTICS_DIR . 'includes/admin/class-wp-statistics-admin-template.php';
 
-        $referrals      = new ReferralsManager();
-        $postsManager   = new PostsManager();
-        $userOnline     = new \WP_STATISTICS\UserOnline();
+        $referrals           = new ReferralsManager();
+        $postsManager        = new PostsManager();
+        $userOnline          = new \WP_STATISTICS\UserOnline();
+        $notificationManager = new NotificationManager();
 
         // Admin classes
         if (is_admin()) {